# api/openapi.yaml
openapi: 3.1.0
info:
  title: DevOps Incident Platform API
  version: 0.0.1
  description: REST API for the DevOps incident management platform.
  license:
    name: MIT
    url: https://opensource.org/license/mit
...
servers:
  - url: http://localhost:4010
    description: Local mock server (pixi run mock-api)
  - url: http://localhost:8080/api/v1
    description: Local gateway and Swagger UI (pixi run compose-up, same paths as Helm)
  - url: /api/v1
    description: API gateway (relative, in-cluster)
paths:
...
  /incidents:
    get:
      operationId: listIncidents
      summary: List incidents with optional filtering
      tags: [incidents]
      security: []
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by incident status
          schema:
            $ref: "#/components/schemas/IncidentStatus"
        - name: severity
          in: query
          required: false
          description: Filter by incident severity
          schema:
            $ref: "#/components/schemas/Severity"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentListResponse"
        "401":
          description: Not authenticated
    post:
      operationId: createIncident
      summary: Create a new incident manually
      tags: [incidents]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentRequest"
      responses:
        "201":
          description: Incident created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
  /incidents/{incidentId}:
    get:
      operationId: getIncident
      summary: Fetch a single incident
      tags: [incidents]
      parameters:
        - $ref: "#/components/parameters/IncidentIdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Incident not found
  /incidents/{incidentId}/description:
    patch:
      operationId: updateIncidentDescription
      summary: Edit the incident description
      description: Set or clear the human-written incident description.
      tags: [incidents]
      security: []
      parameters:
        - $ref: "#/components/parameters/IncidentIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDescriptionRequest"
      responses:
        "200":
          description: Description updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
        "404":
          $ref: "#/components/responses/IncidentNotFound"
  /incidents/{incidentId}/status:
    patch:
      operationId: updateIncidentStatus
      summary: Transition incident status
      description: |
        Any transition between distinct statuses is allowed (including reopening
        a resolved incident, which clears resolvedAt).
        Requires RESPONDER or COMMANDER role.
      tags: [incidents]
      security: []
      parameters:
        - $ref: "#/components/parameters/IncidentIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
        "403":
          description: Insufficient permissions
        "404":
          $ref: "#/components/responses/IncidentNotFound"
  /incidents/{incidentId}/severity:
    patch:
      operationId: escalateIncidentSeverity
      summary: Manually set incident severity
      description: Sets the severity to any level (raise or lower). Requires COMMANDER role.
      tags: [incidents]
      security: []
      parameters:
        - $ref: "#/components/parameters/IncidentIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalateSeverityRequest"
      responses:
        "200":
          description: Severity updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          description: Invalid severity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
        "403":
          description: Insufficient permissions
        "404":
          $ref: "#/components/responses/IncidentNotFound"
  /incidents/{incidentId}/assign:
    patch:
      operationId: assignIncident
      summary: Assign or unassign responders to an incident
      description: Requires COMMANDER or RESPONDER role (can assign themselves).
      tags: [incidents]
      security: []
      parameters:
        - $ref: "#/components/parameters/IncidentIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignIncidentRequest"
      responses:
        "200":
          description: Assignment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
        "403":
          description: Insufficient permissions
        "404":
          $ref: "#/components/responses/IncidentNotFound"
  /incidents/{incidentId}/events:
    get:
      operationId: listIncidentEvents
      summary: List Events from the Event Log for one incident, in chronological order
      tags: [incidents]
      parameters:
        - $ref: "#/components/parameters/IncidentIdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IncidentEvent"
        "404":
          description: Incident not found
  /incidents/{incidentId}/comments:
    get:
      operationId: listComments
      summary: List comments on an incident
      tags: [incidents]
      security: []
      parameters:
        - $ref: "#/components/parameters/IncidentIdParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"
        "401":
          description: Not authenticated
        "404":
          $ref: "#/components/responses/IncidentNotFound"
    post:
      operationId: addComment
      summary: Add a comment to an incident (immutable)
      description: >-
        The comment's authorId is taken from the gateway-injected `X-User-Id`
        header (the validated session identity), never from the request body.
      tags: [incidents]
      security: []
      parameters:
        - $ref: "#/components/parameters/IncidentIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
        "404":
          $ref: "#/components/responses/IncidentNotFound"
...
components:
  parameters:
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    IncidentIdParam:
      name: incidentId
      in: path
      required: true
      description: UUID of the target incident.
      schema:
        type: string
        format: uuid
        example: 018e2c5f-1234-7abc-8def-000000000001
...
  responses:
...
    IncidentNotFound:
      description: No incident exists with the given ID
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Simple error payload for auth and validation failures.
      required: [message]
      additionalProperties: false
      properties:
        message:
          type: string
          example: Invalid email or password
...
    Severity:
      type: string
      enum: [SEV1, SEV2, SEV3, SEV4]
      description: Incident severity. SEV1 is highest impact.
    IncidentStatus:
      type: string
      enum: [open, investigating, resolved]
    Incident:
      type: object
      description: An incident as stored by incident-service.
      required: [id, title, status, severity, createdAt]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Checkout 5xx spike
        description:
          type: [string, "null"]
        status:
          $ref: "#/components/schemas/IncidentStatus"
        severity:
          $ref: "#/components/schemas/Severity"
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: [string, "null"]
          format: date-time
        summary:
          type: [string, "null"]
          description: AI-generated narrative summary. Regenerable on demand.
        summaryGeneratedAt:
          type: [string, "null"]
          format: date-time
          description: When the AI summary was last generated.
        severitySuggestion:
          type: [string, "null"]
          description: 'AI-suggested severity with reasoning, formatted as "SEV<n>: <reason>".'
        severitySuggestionGeneratedAt:
          type: [string, "null"]
          format: date-time
          description: When the AI severity suggestion was last generated.
        solutions:
          type: [string, "null"]
          description: AI-suggested remediation steps, one per line.
        solutionsGeneratedAt:
          type: [string, "null"]
          format: date-time
          description: When the AI solution suggestions were last generated.
        postmortem:
          type: [string, "null"]
          description: AI-drafted postmortem. Only set for resolved incidents.
        postmortemGeneratedAt:
          type: [string, "null"]
          format: date-time
          description: When the AI postmortem was last generated.
        assignedUserIds:
          type: array
          items:
            type: string
            format: uuid
          description: >-
            UUIDs of the responders currently assigned to this incident
            (see PATCH /incidents/{incidentId}/assign).
          example: ["018e2c5f-1234-7abc-8def-0000000000aa"]
    IncidentEvent:
      type: object
      description: One entry from an incident's append-only Event Log.
      required: [timestamp, type, description]
      additionalProperties: false
      properties:
        timestamp:
          type: string
          format: date-time
        type:
          type: string
          example: status_changed
        description:
          type: string
          example: "status: open -> investigating"
        newValue:
          type: string
          description: >-
            New value after the change: the new status for status_changed
            entries and the new severity for severity_changed entries. Lets
            clients color-code timeline entries without parsing the
            description. Absent for other entry types and for events stored
            before this field existed.
          example: investigating
...
    CreateIncidentRequest:
      type: object
      description: Request to create a new incident manually.
      required: [title, severity]
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          example: Database migration rollback needed
        description:
          type: string
          maxLength: 4000
          example: Migration 2026_07_01 left the orders table partially indexed.
        severity:
          $ref: "#/components/schemas/Severity"
    UpdateDescriptionRequest:
      type: object
      description: Request to set or clear the incident description. An empty string clears it.
      required: [description]
      additionalProperties: false
      properties:
        description:
          type: string
          maxLength: 4000
          example: Checkout error rate crossed 5% after deploy v2.4.1.
    UpdateStatusRequest:
      type: object
      description: Request to update incident status.
      required: [status]
      additionalProperties: false
      properties:
        status:
          $ref: "#/components/schemas/IncidentStatus"
    EscalateSeverityRequest:
      type: object
      description: Request to set incident severity (any level, higher or lower).
      required: [severity]
      additionalProperties: false
      properties:
        severity:
          $ref: "#/components/schemas/Severity"
    AssignIncidentRequest:
      type: object
      description: Request to assign or unassign responders.
      required: [userIds]
      additionalProperties: false
      properties:
        userIds:
          type: array
          items:
            type: string
            format: uuid
          description: UUIDs of users to assign. Send empty array to clear all assignments.
          example: ["018e2c5f-1234-7abc-8def-0000000000aa"]
    Comment:
      type: object
      description: An immutable comment on an incident, authored by a user.
      required: [id, incidentId, authorId, text, createdAt]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        incidentId:
          type: string
          format: uuid
        authorId:
          type: string
          format: uuid
          description: UUID of the user who wrote the comment
        text:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateCommentRequest:
      type: object
      description: Request to add a comment to an incident.
      required: [text]
      additionalProperties: false
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 5000
          example: "Rolled back deployment v2.4.1; monitoring error rate now."
    CommentListResponse:
      type: object
      required: [items, total, page, size]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        total:
          type: integer
          minimum: 0
          example: 3
        page:
          type: integer
          minimum: 0
          example: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
          example: 50
    IncidentListResponse:
      type: object
      required: [items, total, page, size]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Incident"
        total:
          type: integer
          minimum: 0
          example: 5
        page:
          type: integer
          minimum: 0
          example: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
          example: 50
...